requirements.txt
numpy
scikit-image
opencv-python
pytest
colorama
tqdm
imagehash
pillow
